feat: move Container Apps Environment to shared stack
Moves Log Analytics Workspace and Container Apps Environment from
per-environment stacks to the shared stack. This works around the
subscription quota limit of 1 Container Apps Environment.
